Remove post-receive-hook, more utils to app
